agents/link-content-scripter/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "link-content-scripter"
version = "0.1.0"
description = "MoFA agent that fetches content from RSS entry links and generates news broadcast scripts"
readme = "README.md"
license = {text = "MIT License"}
dependencies = [
    "numpy<2.0.0",
    "pyarrow>=5.0.0",
    "mofa-core",
    "dora-rs",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]

[project.scripts]
link-content-scripter = "link_content_scripter.main:main"

[tool.setuptools.packages.find]
include = ["link_content_scripter*"]
